+++ src/search/searchcontroller.cpp
@@ -1,3 +1,0 @@
-#include "searchcontroller.h"
-
-SearchController::SearchController() {}